faster fringestopping impl in n2accumulate. memory matters!

+++ lib/utils/CHORDTelescope.hpp
@@ -656,7 +656,7 @@ public:
      *                  first num_dishes elements of this vector.
      **/
     void fringestop_phases_1d(double freq_MHz, const EOP& eop, const EOP& eop0,
-                              std::vector<std::complex<double>>& phases) const;
+                              std::vector<std::complex<float>>& phases) const;
 
     /**
      * @brief   Fill a dishInputFields struct with dish information. Will possibly
